requirements.txt
filehash
pathlib
python-dotenv
requests
schedule
solid-file
rdflib
prettytable
argparse
tornado
webthing

